% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{internal.logfiles}
\alias{internal.logfiles}
\alias{.openlog}
\alias{.addlog}
\alias{.closelog}
\title{Write log files.}
\usage{
.openlog(filename, fun)

.addlog(..., sep = " ")

.closelog(dir.output, cancel = FALSE)
}
\arguments{
\item{filename}{(optional) character. Full name of the file to be produced by the function that calls the internal function. If missing, command \code{get("filename", envir = parent.frame())} is used.}

\item{fun}{(optional) character. Name of the function that opens the log file. If missing, command \code{sys.call(-1)[[1]]} is used, but it won't work when the log is opened inside a foreach loop.}

\item{...}{character. Strings to be concatenated and form a message.}

\item{sep}{a character string to separate the strings passed via \code{...} (see \code{\link{paste}})}

\item{dir.output}{(optional) character vector. Where the outputs of the function that calls the internal functions will be saved.}

\item{cancel}{logical. Cancels the logfile (nothing is produced).}
}
\description{
Internal functions to be called from within another functions (see details).
}
\details{
These functions are used together to write logfiles containing information related
to the outputs produced by the function that calls these internal functions.
The functions should be used as follows:

\code{.openlog} opens a log file

\code{.addlog} optionally adds additional messages to the log.

\code{.closelog} closes the log file opened by \code{.openlog}.

Note that if arguments \code{filename} and \code{dir.output} are missing,
they must be defined in the environment of the function that calls the internal functions (see examples).

The name of the logfile produced is defined by the \code{filename} plus the extension .txt.
Therefore, the logfile has the same name of the product it refers to
(e.g. output_of_a_function.txt).
The log file is saved in directory "logfiles", found inside the folder of \code{filename}.

The content of the logfile is devided in three sections:
\emph{System info}, \emph{Function info}, and \emph{Additional info}.
\emph{System info} includes \emph{User}, \emph{Nodename}, \emph{Operating system},
\emph{System time}, and \emph{logfile version}.
\emph{Function info} includes \emph{logfile function} (i.e. the function that produced the log),
\emph{Data produced}, and \emph{Data location}.
Both \emph{System info} and \emph{Function info} are automatically produced by \code{.openlog}.
\emph{Additional info} includes all the messages added with function
\code{.addlog}, plus the time elapsed (in seconds), which is automatically added by \code{.closelog}.

\code{.addlog} may be used as many times as needed;
each message added will correspond to a line in the logfile.
}
\examples{
x<-file.path(tempdir(),"S2A_L2A_20171002_T29SNC.tif")
y<-dirname(x)
fun<-function(a,b){
  logfile:::.openlog(a)
  logfile:::.addlog("Add", "a message", "to the logfile")
  logfile:::.closelog(b)
}
fun(x,y) # check logfile in folder y

fun<-function(filename,dir.output){
  logfile:::.openlog()         # no argument is provided as defined in fun
  logfile:::.addlog("Add", "a message", "to the logfile")
  logfile:::.closelog()        # no arguments are provided as defined in fun
}
fun(x,y)

}
\keyword{internal}
